--- app/src/main/java/com/example/resona/data/remote/model/LoginResult.kt
@@ -1,5 +1,0 @@
-package com.example.resona.data.remote.model
-
-data class LoginResult(
-    val isNewUser: Boolean
-)
