+++ app/src/main/java/com/example/resona/ui/SplashFragment.kt
@@ -1,4 +1,4 @@
-package com.example.resona
+package com.example.resona.ui
 
 import android.os.Bundle
 import android.view.View
@@ -6,16 +6,17 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavOptions
 import androidx.navigation.fragment.findNavController
+import com.example.resona.R
 import com.example.resona.data.local.TokenManager
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @AndroidEntryPoint
 class SplashFragment : Fragment(R.layout.fragment_splash) {
-    @Inject lateinit var tokenManager: TokenManager
+    @Inject
+    lateinit var tokenManager: TokenManager
 
     override fun onViewCreated(view : View, savedInstanceState: Bundle?){
         super.onViewCreated(view, savedInstanceState)
