+++ app/src/main/java/com/example/resona/ui/login/viewmodel/LoginViewModel.kt
@@ -1,14 +1,14 @@
-package com.example.resona.data.remote.model
+package com.example.resona.ui.login.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.resona.data.local.TokenManager // 1. TokenManager 임포트
+import com.example.resona.data.event.Event
+import com.example.resona.data.local.TokenManager
 import com.example.resona.data.remote.api.KakaoLoginResponse
 import com.example.resona.data.repository.AuthRepository
-import com.example.resona.data.event.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
