+++ app/src/main/java/com/example/resona/ui/onboarding/fragment/OnboardingRecommendFragment.kt
@@ -1,4 +1,4 @@
-package com.example.resona
+package com.example.resona.ui.onboarding.fragment
 
 import android.os.Bundle
 import android.text.Editable
@@ -9,10 +9,10 @@ import android.widget.EditText
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
-import com.example.resona.data.remote.model.OnboardingViewModel
+import com.example.resona.R
+import com.example.resona.ui.onboarding.viewmodel.OnboardingViewModel
 import com.example.resona.databinding.FragmentOnboardingRecommendBinding
 import com.example.resona.ui.main.MainActivity
-import kotlin.getValue
 
 class OnboardingRecommendFragment : Fragment(R.layout.fragment_onboarding_recommend) {
     private var _binding: FragmentOnboardingRecommendBinding? = null
