+++ app/src/main/java/com/example/resona/ui/onboarding/viewmodel/NicknameViewModel.kt
@@ -1,10 +1,8 @@
-package com.example.resona.data.remote.model
+package com.example.resona.ui.onboarding.viewmodel
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.resona.data.remote.model.NicknameRequest
+import com.example.resona.data.remote.model.ApiResult
 import com.example.resona.data.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
