+++ app/src/main/java/com/example/resona/ui/onboarding/viewmodel/ProfileImageViewModel.kt
@@ -1,11 +1,13 @@
-package com.example.resona.data.remote.model
+package com.example.resona.ui.onboarding.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.resona.data.repository.MainRepository
 import com.example.resona.data.event.Event
+import com.example.resona.data.remote.model.ApiResult
+import com.example.resona.data.remote.model.ProfileImageResponse
+import com.example.resona.data.repository.MainRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
